SecureKeyStorage protocol in its own file
